+++ django_boost/admin_tools/apps.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import warnings
+
 from django.apps import AppConfig
 
 
@@ -7,5 +9,9 @@ class AdminToolsConfig(AppConfig):
     name = 'django_boost.admin_tools'
 
     def ready(self):
-        # Implement your startup processing
-        pass
+        warnings.warn(
+            "'django_boost.admin_tools' is deprecated and will be removed in "
+            "django-boost 4.0; use 'django_boost.contrib.admin_tools' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )